mbox-prepare: add interactive mode
In this mode we will fire up $EDITOR for the user to manually fix the
subject to be re-parsed by git-mbox-prepare. This allows quick
workarounds without editing the mbox file by hand. We can teach
git-mbox-prepare some tricks later to deal automatically with the common
cases, but for now the interactive mode should suffice.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>