git-pile: implement format-patch
This uses the approach of checking out the pile branch in a temporary
checkout so we can write the diff of the git tree. The other approach
would be to use `git diff --index`, treating them basically as a dir
diff.  This creates problems with files/dirs we don't want to include in
the diff (e.g. .git, .gitignore, etc) and git ends up generating
absolute paths in the diff in some circustances. It's also not
consistent with the use of /dev/null when a file is deleted/added: in
the "diff --git " line it doesn't use /dev/null, while in the "---
old-file\n+++ new-file" it does. The end result is that we need to keep
patching the diff.

With this approach we also don't need to parse any diff, but just reuse
the one producedby `git diff -p --raw`. Bonus point that if for any
reason the pile dir was dirty, it will continue to work as intended.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>