git-pile: fix format-patch using wrong subjectprefix
If the user has a format.subjectprefix set in the repo, he may screw the
patches generated with genpatches because we were not enforcing it. We
now enforce a generic "PATCH" prefix for the patches stored in the pile
branch, but allow the user to override it when calling
"git-pile format-patch" since a mailing list may be shared with more
projects.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>