build: git-pile: include order file in data dir
I'm not sure the MANIFEST.in file is needed. It seems redundant and on
my local system it works without it. However it is the documented method
at https://python-packaging.readthedocs.io/en/latest/non-code-files.html
and I also have it working like this in codespell.  At the risk of just
propagating the mistake, do the same here.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>