git-pile: print help when called outside of git directory
The default argument to a function is called when the function is
defined, not when it's called. Just move the default to the caller so we
can at least show the help message outside of a git directory.

[lucas@lucasdemarchi-desk tmp]$ git pile
fatal: not a git repository (or any parent up to mount point /)
Stopping at filesystem boundary (GIT_DISCOVERY_ACROSS_FILESYSTEM not set).
Traceback (most recent call last):
  File "/home/lucas/p/git-pile//git-pile", line 5, in <module>
    import git_pile.git_pile as git_pile
  File "/home/lucas/p/git-pile/git_pile/git_pile.py", line 112, in <module>
    def temporary_worktree(commit, dir=git_root(), prefix=".git-pile-worktree"):
  File "/home/lucas/p/git-pile/git_pile/git_pile.py", line 70, in git_root
    return git("rev-parse --show-toplevel").stdout.strip("\n")
  File "/home/lucas/p/git-pile/git_pile/helpers.py", line 49, in __call__
    ret = subprocess.run([self.cmd] + l, *args, **kwargs)
  File "/usr/lib64/python3.6/subprocess.py", line 418, in run
    output=stdout, stderr=stderr)
subprocess.CalledProcessError: Command '['git', 'rev-parse', '--show-toplevel']' returned non-zero exit status 128.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>