git-pile: change default pile directory
Change it to patches/ to be more compatible with default value from
quilt and probably what people may have in their .gitignore.  My
original intention with having a different directory was so I could keep
it working on a repository that had qf already configured. Admittedly it
was not necessary as in this specific repository I can deviate from the
default.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>