git-pile: make git_root() return outermost git directory
git-pile works by using git-worktree for the pile branch and for some
operations. However "git rev-parse --show-toplevel" doesn't return the
topmost directory, which means that if the user is inside one of these
directories (e.g. patches/) and calls git-pile we will try to work
relative to the wrong dir.

Fix it by first checking what's the git-common-dir.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>