git-pile: implement setup() command
This contains many more sanity checks than in other commands in the hope
it won't fail misteriously for people and they will give up on the new
tool. After all this is likely to be the first command they will run.

The idea here is that the user will pass 2 branches: the pile, and where
the result should be saved. This reinforces the idea of how git-pile
should be seen by the end user:

                   .-------> genpatches ---------.
                   :                             v
             RESULT_BRANCH                  PILE_BRANCH
                   ^                             :
                   '-------- genbranch <---------'

So, in the end the user knows mostly about the two branches and how to
convert from one to the other. With these two arguments we derive
the options that we save in our config.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>