mbox-prepare: copy git-format-patch's impl for subject parsing
Instead of indefinitely trying to blacklist some chars and using them
in a call to transliterate(), just copy what git is doing. We are now
doing more a whitelist than a blacklist, like git does, with some
extra logic to skip duplicates and strip the result.

This already fixes 2 cases:
1) We can have a dot in the middle of the subject
2) "'" char should be blacklisted

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>