git-pile: workaround git-worktree without pruning
When people play with simply removing the directory of a worktree
instead of calling `git worktree remove`, the result from
`git worktree list` can get out of sync with reality. We could force a
`prune` when we are listing for the first time, but we may as well just
check if the directory exists since that's what we are after.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>