git-pile: preserve index line for binary diffs
When we have a binary diff, we must preserve the index line, otherwise
it's not possible to regenerate the tree. While trying to apply the
patch with the genbranch command git will spit:

error: cannot apply binary patch to 'bla.bin' without full index line

We don't pass a `--binary` to git-format-patch since we assume the
branch will be generated on the same repository that generated the
patches hence it will have the git objects. It's not always true for the
git-pile usecases, but the use of binary patches is also not common.
This can be improved later.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>