helpers: use universal_newlines rather than specific encoding
We don't care much about the encoding since most of the commands are
callouts to git.  What we care here is that it's opened as text rather
than as binary.

encoding and text kwarg were only added in python 3.6 and 3.7
respectively. So use universal_newlines from older versions (still
compatible with the new ones) to force the stderr/stdout to be in text
mode.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>