git-pile: allow to generate filename from single commit
When we want to know the filename that will be on "series" for a single
patch, we want to reuse the same function so they always match and we
don't risk getting different places out of sync. So, allow to generate
the filename for a single commit as well: this will be used on
subsequent patches.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>