git-pile: format-patch: filter diff in coverletter
Now the diff we generate and append to the coverletter will contain
only the patches reported by git-range-diff as changed, added or
removed.  This way the overall diff matches the patches we are sending
in the series, and applying it with "git am 0000-coverletter.patch"
will more likely apply without conflicts.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>