git-pile: Include pile-commit in the cover letter
Like the base-commit info added by "git format-patch --base", but here
in the context of git-pile: it records the base commit in the pile
branch this applies to. With that it's easier to reproduce the exact
same branch as the person submitting the series had, and eventually
rebase this branch rather than always trying to apply on top and risking
conflicts.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>