git-pile: fix patch order
When we moved to use git-range-diff in format-patch in 82708d3
("git-pile: implement format-patch based on range-diff") we removed the
sorting of the patches. This results in wrong ordering of the series
with the added patches to the end of the series, although the cover
letter is still correct.

Explicitly sort the list with added and changed patches when creating
them.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>