git-pile: Add initial version of am command
This allows to easily applies patches to the pile branch that were
generate with git-pile format-patch. We parse the mbox file passed as
argument and assume the first patch is the cover.

We also parse the cover letter to make sure it's generated by git-pile.
The following fields are expected to be in the cover, that matches what
v0.9 generates:

    * signature should be in the form of "git-pile .*"
    * content between commit message and diff must contain:
        - baseline commit
        - pile commit

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>