Improve error message when applying a patch
People get scared when python explodes with a backtrace, so improve the
error message.

This happens when the baseline doesn't exist in your local repository.
This may be because there's garbage in the baseline line of the patch
you are trying to apply, or because you simply need a git fetch.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>