Fail format-patch if the pile and the branches are not in sync
If the branches we're diffing to generate the pile patch do not
contain the specified or inferred baseline, git range-diff can
detect unrelated patches in the diff.

Instead of generating an incorrect pile patch, throw an error
with an explanation of what went wrong.

v2: keep supporting the possibility of having 2 different baselines (Lucas)

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>