git-pile: workaround bad email clients
When Thunderbird saves the email as mbox, it uses DOS line endings. If
the file was passed directly to git, it would be smart enough to figure
everything out. However, `git-pile am` had to decode it to give some
better feedback to the user:

1) separate the cover from the rest - intention is for it to work with
patchwork without the user needing to split it
2) catch some common mistakes like not giving a cover letter as
argument, or a mbox with the wrong order, etc, etc

After handling it, mailbox encodes it in base64, which apparently
confuses git on handling the dos line endings (mixed with the unix line
endings from the header). Handle it by making the dump() method smarter,
printing all the information we need rather than just trying to convert
the message object to string.

This also handles the case of a missing "From" as the first line, which
also happens when someonone tries to pipe the email from mutt rather
than saving it.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>