git-pile: format-patch: glob the filter
When we have all patches with the same subject, the file names provided
by `git log --format=%f` are all the same. In `genpatches` we make them
unique by going through the list of files and disambiguating them by
adding a number suffix. This works well in the case of genpatches, in
which we have all the patches we possibly want.

However, for format-patch we reuse that logic to create a filter from
the output of range-diff. We need this filter because the diff to attach
to the coverletter would be otherwise too noisy due to all the hunk diff
changes. However there's no direct mapping from the file name generated
by `git log --format=%f` to the file name actually stored in the pile.
As a result, if commits have the same name we will fail to apply the
changes to some of them because the file names won't match.

This changes generate_series_list() to accept the suffix as a param and
use that in format-patch. We now give it a glob "*.patch" so the diff in
the cover letter will include any change to any patch with the same
name. This potentially increases the noise, but should be minimal as
patches with the same subject line are usually very few.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>