git-pile: decode headers
We need to decode the header to get the bytes representation and then
try to use the charset returned to try to decode with that charset.
I don't know why, but that is failing for me with cover generated by
José. The charset returned is unknown-8-bit, that then fails to decode
the bytes returned: b'Jos\xc3\xa9 Roberto de Souza <jose.souza@intel.com>'

Since our cover is always generated as UTF-8, try hard to workaround it
by forcing utf-8 as charset in case the reported charset by the decoder
failed.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>