git-pile: teach "am" to workaround broken mbox
If the mbox doesn't start with a line "From ...." mailbox.mbox() would
create an empty mbox with 0 messages. When we pipe an email from
mutt/neomutt and possibly other email clients, this line is omitted.
Then when people try to apply a patch with `git pile am`, they would
have to add that line back, or ensure they are exported by their email
client in the right way.

Move the decision "from stdin or from file" back to PileCover.parse()
and make them compatible to lookahead on first line. We now always stash
the content in a temporary file, possibly adding an additional line with
the "From ..."  line.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>