git-pile: rename temporary worktree
git 2.23 started mangling the name of the worktree to avoid having
invalid refnames. As part of this sanitize part, '.' is transliterated
to '-'. As a result the temporary worktree we created for generating a
branch has a different git-dir than the name of the directory.

Avoid the issue completely by getting rid of the dot in the temporary
worktree name. It's not that important for us, and if we leave worktrees
around due to failed commands, it's better be noticeable by the user.

See commit 1de16aecf51d ("worktree add: sanitize worktree names")
in git.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>