git-pile: allow format-patch to add header
This wraps the format-patch's --add-header option. The idea is that it
should be simple for one to create an additional header to track patches
that are part of a pile, or part of an organization. Example scenarios:

X-Pile: my-awesome-project

or

Organization: Intel

Any for these should be set in pile.format-add-header git configuration
in order to be used.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>