git-pile: format-patch: append full diff
When generating the patches to send, also generating a last patch that
is only intended for review. This patch is the diff between the current
branch and the upstream one.

This applies only when generating a simple format-patch diff in which
the baselines match. If the baselines don't match, the diff could be
very big and outside of the changes done by the patch author, so we just
disable it in this case.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>