genbranch: add option to just apply the patches
Sometimes we only want to know if the patches apply and want to check
the final result. For those cases we may just apply the patches rather
than have all the commits generated.

Not so creative for names, so make it "--dirty" since it will leave the
user the git tree dirty.
I added some:

Time to generate the dirty result with --in-place, too: 0m0.985s

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>