git-pile: am: Correctly decode From line
The current code was not appending the email address of the From line
if it needed decode. Causing "git am" to fail with "empty ident name
(for <>) not allowed".

A simple way to test is replace the From line of any git-pile file
with my From line:
From: =?UTF-8?q?Jos=C3=A9=20Roberto=20de=20Souza?= <jose.souza@intel.com>

Signed-off-by: José Roberto de Souza <jose.souza@intel.com>