Stop using PILE_RESULT_HEAD
We have some problems with the way we handle PILE_RESULT_HEAD:

a) people are not used to have a ref like and the error message seems
crypt if they don't mentally map the PILE_RESULT_HEAD is like "HEAD" or
"FETCH_HEAD" and I can reset --hard to it

b) older git versions (at least <= 2.22) have problems in finding heads
saved on other worktree directories. This means that
`git rev-parse PILE_RESULT_HEAD` in a different directory will not
return the right commit. If people did the genbranch from the patches
directory and then go to the main directory and try to reset, it won't
work. Recent versions of git have that fixed, but we'd better avoid that
issue.

So now we only print the long commit hash. That will be valid until
`git prune` is called which should be sufficient and hopefully trigger
the mental mapping of "this is a commit I can reset my branch to".

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>