am: check pile-commit is available on remote
When we call `git pile format-patch` it will use the pile branch to
calculate the diff that goes on the cover-letter. However this only
makes sense outside a local machine if who is receiving that
cover-letter can checkout the pile-commit. When generating a v2
of a patch series, very often we do:

	git pile am <v1-cover.mbx>
	# .. fix conflicts
	git pile genbranch -i
	# .. fix conflicts
	# .. review feedback, other changes, etc

Then if we forget to do `git -C patches reset --hard HEAD^`, the next
`git pile format-patch` will be the diff between v1 and v2, not the diff
between the current (remote) state of the pile and v2.

To fix that, check the pile-commit in the local pile branch is ancestor
of the pile commit in the remote pile branch. If needed we may add an
argument to force bypassing this check in future.

Fix: #18

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>