Dot not try to split mbox
Just open it as a single message. Earlier we needed to support splitting
an mbox because the only interface Patchwork would give us was to
download the patch series with cover. Now patchwork has support to
download only the cover, so we don't have to try to split the mbox
anymore.

This fixes an issue with mailbox.mbox() erroneously splitting the mbox
in the first occurrence of "^From" in the body of the message.

Fix: #19

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>