genbranch: support config for --committer-date-is-author-date
Add config to use only the author date and make it the default. The
reasoning is that in a branch maintained with git-pile, the committer
date will always be the last time that commit was generated and that
will change every time the pile is generated. The time basically doesn't
add any value.

In order to support workflows where we want the same commit SHAs to be
generated if the same pile + baseline is used, let's switch to this by
default.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>