Add helper for orderedset
Second time, in different projects, that I was caught by this: a set in
python is unordered and using it to remove duplicates may change the
order. Add a 'orderedset' helper which uses a dict to mimic an ordered
set. This allows us to remove duplicates, but guarantee the ordere
remains the same.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>