format-patch: allow to reuse cover message from another commit
When we are generating new versions of a patch series, a common workflow
is to apply the previous patch with:

	git pile am cover.mbx
	git pile genbranch -i

Then when generating the new version:

	( cd patches && git reset --hard HEAD^ )
	git pile format-patch --subject-prefix="PATCH v2"

In this scenario, we then have to copy and paste the previous message in
the cover and modify it to take into account the changes. With this
additional argument, we can point git-pile format-patch to the previous
commit in the pile branch to fill out the cover-letter:

	( cd patches && git reset --hard HEAD^ )
	git pile format-patch --subject-prefix="PATCH v2" -C <commit>

<commit> can easily be found by simply printing it in the patches
directory before resetting or just checking on git-reflog.

This is inspired by the same option, -C / --reuse-message, that is found
on git-commit (although git-format-patch doesn't have such an option).