format-patch: Set the REVIEW patch content type to text/x-diff
If the end result is empty (just reordering patches etc), the REVIEW
patch will be an empty patch. Patchwork doesn't parse empty patches as
patches if the content-type is text/plain, but using text/x-diff
forces it to grab the whole email as the patch. We don't need to care
what ends up in the patch as far as patchwork is concerned, as no one
needs to feed that patch as such back to git, only the cover letter is
used.

Signed-off-by: Petri Latvala <petri.latvala@intel.com>