init: Give an explicit branch name on git-init
git >= 2.28 has a setting that the user may have set to change the
default branch used by git-init. This was due to projects moving out of
"master" branch and calling it something else, like "main".

When we do a temporary repo to initialize an orphan pile branch, let's
give "pile" as the branch name explicitly: "master" may not be the
default branch and we would then fail when we fetch the temporary repo.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>