am: prompt user about auto-solve if --fuzzy it not specified
When the user forgot to pass the --fuzzy, instead of just bailing out,
add a prompt asking if we should try to auto-solve the conflicts.
It should be a better experience than having to run `git am --abort`
and `git pile am --fuzzy` again.

Also make sure we only prompt anything if running from terminal,
otherwise we remain with "no" as default for compatibility with previous
versions.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>