Plumb genlinear-branch command
Add empty new command to generate a linear branch off the pile branch.
Intention here is to create a branch with the result of
`git pile genbranch` for each revision in the pile. This allows to
visualize exactly what was the "overall change" in each revision without
having to read the interdiff in the pile. This is the same as if we were
to apply the "REVIEW" patch generated by `git pile format-patch` for
each pile commit.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>