genlinear-branch: basic implementation for new command
Implement command to generate a linear branch from the current pile log.
This calls genbranch for each commit in the pile. The result for each of
them is used to handcraft a branch. The branch needs to be saved in the
config file or passed in the command line.

This is the equivalent of applying the "REVIEW" patches for each
git-pile formatted series. Only the net changes survive, with the others
captured as empty commits.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>