genlinear-branch: add incremental mode and make it the default
Since calling genbranch for each rev in the pile branch can be very
expensive, it's not expected that we would redo it every time. Instead
allow use to figure out from the linearized branch what are the revs
still missing. This is done by looking at the git-notes we save and
checking for a "pile-commit:" entry. Then we only iterate over the
missing commits. This is the default behavior, can be overriden by
--no-incremental/--recreate argument.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>