genlinear-branch: handle keyboard interrupt
In order to allow ctrl-c to work as expected we need to do a few
changes:

1) Start the long running git-am process in its own session group.
Otherwise the signal is also sent to the child process. If the child
fails, we may not get the KeyboardInterrupt exception correctly in
git-pile

2) handle KeyboardInterrupt and raise it: this will allow the "finally"
code blocks to run, cleaning up the temporary checkout. The exception
will be handled by main()

3) Although it costs a little bit per iteration, we also call update-ref
every time instead of just at the end. We could move it to a finally
code block, but I feel that for a long-running process it's good to see
the branch being updated in another shell.

With this change, we can leave it running for some hours, stop and
continue later.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>