genlinear-branch: switch to git-pile-genlinear-branch notes ref
The notes are added to the commits in the separate linearized
branch, so initially I thought just using the default refs/notes/commits
ref would be good so it's automatically displayed in git-log/git-show.
genlinear-branch would be the only command writting notes to those
commits, so even overwritting them would not be a problem.

However, since ideally the notes would also be shared in a remote
repo, there is no way to separate the notes: we'd be pushing every note,
including from other branches (see `git log refs/notes/commits`). If the
user use notes for something else, they would also be pushing their
(private) notes to the server.

So make sure we use our own ref for our annotations rather than
sharing the default. The inconvenience may be overcome by either
passing --notes=git-pile-genlinear-branch to git-log/git-show
or to set it in the config core.notesRef.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>