genbranch: enable split index
As found out by Matt Roper <matthew.d.roper@intel.com>, we can improve
a lot the time taken by git pile genbranch by using the core.splitIndex
if we are using a big repo like the Linux kernel.

Initially we enabled this in the repo as a configuration, however
leaving it enabled creates some problems during rebase operations:
git leaves behind the .git/sharedindex.* files.

So, instead of enabling this per repository, let's enable it only during
the git-am operation inside the genbranch command.

With this change a genbranch in a large repository like the Linux kernel
and around 2200 commits take ~30s instead of 180s, so we have a ~6x
performance improvement.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>