genpatches: speedup patch generation
This speeds up the genpatch() implementation (and by consequence the
format-patch subcommand). We don't need to fork one git-format-patch for
each commit. We can just operate on the range with a single fork and
later fixup the filename, just like we fixup the Subject line.

On a big repository like the Linux kernel, with ~2300 patches managed by
the pile and a non-trivial change, these are the numbers:

Before: 3m35.559s
After: 33.302s

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>