am: protect against 2 calls to am
Previously we would just rely on git-am failing if called 2 times in
row. However since now we have a fallback, this would erroneously
trigger the fuzzy application *for the previous patch*, not for the
mbox the user is passing. On a second call we can't distinguish git-am
failures, so just add an extra check.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>