genlinear-branch: use the branch name to store notes
It's better to name the notes ref after the branch name we are using
rather than the tool we are using. This also allows us to avoid
conflicts if there are more piles in the same repo: we don't want a
single notes branch to store all the notes for all the branches.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>