genlinear-branch: skip notes checking when unavailable
If we are calling the command for the first time, the notes won't be
created yet, so there is nothing to check. It wouldn't be so bad if we
didn't loop through all the pile to check if any of the commits have
a note, making git spit several warnings like

	warning: notes ref refs/notes/bla is invalid

Just return earlier in this case.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>