Override whitespace config
Besides the warnings we see when applying the interdiff, if the user has
configured apply.whitespace=strip, the final patch saved in the pile
will be corrupted. This causes a subsequent genbranch to fail due to
malformed patch. Critical here is to override the config during
git-pile-am phase, but to make sure the result branch is the same
regardless of configuration, let's override it in genbranch too, but
this time with "=warn", so we are at least aware there may be issues
with the patch.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>