format: Add --compose|--edit
Allow to open the editor to modify the cover letter as soon as
format-patch returns succesfully. This can be either set in the command
line with --[no-]compose|--[no-]edit or set in the configuration with
`git config pile.format-compose true`, to make it the default behavior.

If no configured and not set in the command line, default is not to open
the editor.

Editor chosen follows the same logic git uses: see git-var(1).

Fix: #63
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>