format-patch: add --no-range-diff-filter
Allow th user to override format-patch behavior of filtering the cover letter by
the list of patches that appear in the git-range-diff output.

Do not filter out patches that have changes on diff hunk numbers only.  To
reduce the diff and possible conflicts with other patch series, git-pile by
default filters out changes to patches that only change the line numbers as
recorded in the diff hunk headers.

For some rare cases in which a patch to a file removes a large number of lines,
and another patch on top changes lines nearby, this may fail. Passing
--no-range-diff-filter option allows git-pile to output the full diff in the
cover letter, as if git-pile genpatches had been called, avoiding the issue.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>