am: re-open tty after reading pipe
When we are piping a patch to git-pile am, we read it until EOF.  After
that we don't want to read anything from the pipe anymore (and the
sender process isn't supposed to send anything). So, make git-pile
re-open /dev/tty so we can still prompt the user when/if needed.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>