genlinear-branch: Add argument for hook before/after genbranch
It may be desired to run a command just before/after generating the branch.
For example this can be used to reset the branch to a partial state, or
remove something from the series file, or log something.

Both pre and post commands are executed with the pile as CWD, with the
first (and only) argument the location of the result branch.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>