genbranch: Support caching
Save time and power for users working with a very long series (talking
about thousands of patches here).

An important missing feature here is cache pruning, which should be done
in the near future.

The table below summarizes timing information for an experiment using
this implementation with a pile containing a little more than 3600
patches.

    ======================================================
    Offset     :1:     :2:     :3:     :4:     :5:     avg
    ------------------------------------------------------
        0%  41.427  41.587  41.157  41.228  40.997  41.279
       25%  30.718  30.315  30.886  30.930  30.893  30.749
       50%  19.827  19.937  19.976  20.375  19.799  19.983
       60%  16.000  15.991  16.071  16.090  16.306  16.091
       70%  12.321  12.411  12.316  12.466  12.327  12.368
       80%   8.445   8.327   8.371   8.282   8.432   8.372
       90%   4.651   4.671   4.689   4.716   4.701   4.686
      100%    .805    .797    .799    .808    .799    .802
    ======================================================

The experiment first makes a first run of genbranch to create the cache
file, saving it as the "original cache". Then, for each run, it:

 1) resets the result branch and the pile to its original state and
    replaces the current cache with the "original cache". The cache
    needs to be reset because otherwise columns :2: through :5: would
    all have timings similar to the case of offset=100%.

 2) updates the pile by modifying the "Subject:" line of one patch in the
    series in order to cause the creation of new commits from that point
    forward. The column "Offset" represents the location relative to the
    size of the series where the patch is modified (e.g. 0% means that
    the first patch is modified; 50%, the patch in the middle).

 3) runs genbranch.

Each column ":i:" represents the i-th run for the given offset. The
values for those columns are execution time in seconds. The last column
is the average of those times.

It's worth noting that the first row (offset=0%) is equivalent to a
genbranch without using the cache, since all patches are applied; and
that the last row (offset=100%) is equivalent to the case when there
have been no modifications in the pile (or that any modification is
already cached, e.g. when using a previous revision of the pile already
in the cache).

Signed-off-by: Gustavo Sousa <gustavo.sousa@intel.com>