test: Add test using alternate directory
The `genpatches` subcommand provides a `-o` option to allow outputting
the result in directory different from the configured one. Add a test
allowing to check that `genbranch` works using the generated result.
Currently `genbranch` does not allow us to pass the input directory as
an option, so we resort to overriding the `pile.dir` configuration.
A command line option should be added in a future patch, which should
come with the necessary changes to this test.

Signed-off-by: Gustavo Sousa <gustavo.sousa@intel.com>