test: Add support for using coverage
Implement this in non-intrusive way by defining a shim script that wraps
calls to git-pile with `coverage run`. Coverage data can be collected
with:

  COVERAGE=1 bats test

Then the command `coverage` can be used to generate reports.

NOTE: Another way to implement this would be changing the git-pile
script to use coverage API[1] directly. That would result in adding code
that would not be used by the end-user and would require some
workarounds, like encapsulating the main logic (including the `import
git_pile.git_pile as git_pile` statement) into a function, which would
have to be called after we tell coverage to start recording data.
Because of that, prefer to use the shim-based approach.

[1] https://coverage.readthedocs.io/en/6.5.0/api.html#api

Signed-off-by: Gustavo Sousa <gustavo.sousa@intel.com>