test: Define and use common functions for caching tests
Checks for the output are really "boileplate" code in the sense that we
want to check if the output of genbranch matches the `n - k` commits of
the generated branch, with `n` being the number of commits from the
baseline and `k` the number of patches skipped by genbranch because of
the cache. As such, add common functions to call genbranch and to assert
cache usage.

The function `run_genbranch` will be modified later to add assertions
common to almost all calls.

Signed-off-by: Gustavo Sousa <gustavo.sousa@intel.com>