Add worktree-aware pile
This allows to have multiple piles in the same repository when
configured with extensions.worktreeConfig in git. Basically this adds
another config file per worktree in $GIT_DIR/config.worktree.

Note that all read operations we want to do without limiting the search
scope for the configuration, i.e. without passing --worktree. Also, if
extensions.worktreeConfig is not set, the user is expected to be using
just one pile checkout and we can't pass --worktree to any command.
The write operations are steered towards the worktree config: those are
mainly used when initializing a new pile via init/setup and a few
other other corner cases.

This preserves the intended behavior in git: settings in config.worktree
have preference over the repo, system, or global configuration, but we
should still read those so the user has a chance to keep common things
in one of them rather than per worktree.

Fix: #7

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>