Raise minimum requirement for git
Now that we have worktree-aware configuration, it's time to raise
minimum git requirement. The worktree support was added in git 2.20.
Currently we report the minimum version is 2.19 due to the addition
of git-range-diff. Raising to 2.20 shouldn't be a problem for "out of
the box" support for distros because they either package git < 2.19
or git >= 2.20 (confirmed with with Ubuntu, Fedora, RHEL/Centos and
Debian).

Checking for each required feature in runtime is troublesome as we
add more requirements and some git versions change their behavior
slightly. Just parse the output from --version and use that.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>