gitutil: Add module and move stuff inside
Let's move git helpers that are not dependent on specific pile command
arguments to a separate module. While at it, rename some of them so that
all helpers are prefixed with "git_".

This has the advantage of allowing those helpers to be used in other
modules and it also reduces a bit the number of lines in git_pile.py.

Signed-off-by: Gustavo Sousa <gustavo.sousa@intel.com>