Make git-worktree add compatible with git 2.25
git worktree add was failing in Ubuntu 20.04 that has git 2.25.x.
For some reason it doesn't like the patches directory not ending in a
slash. Although this could be fixed on the user side by forcing the
slash in the config, there is no harm in adding one directly in
git-pile.

This was the only fix needed for running the complete testsuite in that
distro under git 2.25. After the fix:

	28 tests, 0 failures

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>