genbranch: Add --external-pile option
Right now we can override the git config directly to make the pile dir
temporarily point to an arbitrary location. However, as we want to be
able to run git-pile command from both the pile and result-branch
checkouts, this won't be possible anymore: we will rely on the worktree
list to find what checkout we have.

Add an option --external-pile for genbranch so we can still use
arbitrary directories if needed.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>