test: Move add_pile_commits/setup_second_pile to common
Move the helper functions to common.bash so they can be shared with
multiple tests.

One thing not entirely obvious is that for using functions from
common.bash inside the tests, the file needs to be loaded again in the
setup() function - loading it in the setup_file is not sufficient.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>